Split typechecker & evaluator tests
