+++ library/Core/test/Main.hs
@@ -3,17 +3,19 @@ module Main where
 import Common.Context (contextTests)
 import Conv (coreConversions)
 import Erasure (erasureTests)
+import Evaluator (evaluatorTests)
 import qualified IR.Weak as Weak
 import Juvix.Library (IO)
 import Parser (coreParser)
 import qualified Test.Tasty as T
-import Typechecker (coreCheckerEval)
+import Typechecker (typecheckerTests)
 
 coreTests :: T.TestTree
 coreTests =
   T.testGroup
     "Core tests"
-    [ coreCheckerEval,
+    [ typecheckerTests,
+      evaluatorTests,
       coreConversions,
       coreParser
     ]
